Implement JSON as an adapter on the PackageMetadata object, simplifying the experience for the consumer.
